2010-01-15 20:48 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)  * harbour/contrib/hbnetio/netiosrv.c     ! fixed c&p typo - it should fix problem reported by Viktor in netiostv
  * harbour/contrib/hbwin/win_prn1.c
    * casting

+++ harbour/contrib/hbwin/win_prn1.c
@@ -288,7 +288,7 @@ HB_FUNC( WIN_TEXTOUT )
 
          if( iWidth < 0 && nLen < 1024 )
          {
-            int n = nLen, aFixed[ 1024 ];
+            int n = ( int ) nLen, aFixed[ 1024 ];
 
             iWidth = -iWidth;
